requirements.txt
# Python 依赖
pytest>=7.0.0
playwright>=1.40.0
pytest-playwright>=0.4.0
pytest-html>=4.0.0
pytest-xdist>=3.0.0
python-dotenv>=1.0.0
